+++ ruma-push-gateway-api/src/push.rs
@@ -1,3 +1,0 @@
-//! Endpoints for `HTTP` pushers to send notifications of events to Push Gateways.
-
-pub mod send_event_notification;