+++ ruma-push-gateway-api/CHANGELOG.md
@@ -2,6 +2,4 @@
 
 Improvements:
 
-* Add endpoints:
-
-  `push::send_event_notification::v1`
+* Add `send_event_notification::v1` endpoint